MERN app with Chakra UI frontend

+++ Backend/controllers/product.controller.js
@@ -31,12 +31,17 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async(req, res) => {
     const {id} = req.params
+
+    if(!product.name || !product.price || !product.image) {
+        return res.status(400).json({success:false, message: "Please provide all fields"});
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted"});
     } catch (error) {
         console.log("Error deleting product:", error.message);
-        res.status(404).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server error"});
         
     }
 };
